scripts/requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
pdfplumber==0.11.0
python-dateutil==2.9.0
requests==2.31.0
